version 0.6.0: update to latest Figshare API, compile with Java 17, use parent pom
* remove `FigshareCategory.UNCATEGORIZED` static example category, as there is no such category in latest Figshare API.
  For testing, use newly added `FigshareCategory.SOFTWARE_TESTING`
* `ArticlePost.categories` property is now a `List<Long>`, rather than `List<Integer>`
* use parent pom, compile with java 17

+++ src/main/java/com/researchspace/figshare/model/ArticlePost.java
@@ -18,17 +18,17 @@ import lombok.Singular;
 @NoArgsConstructor
 @Builder()
 public class ArticlePost {
-	
+
 	String title, description;
-	
+
 	private @Singular List<Author> authors;
-	
-	private @Singular List<Integer>categories;
-	//required so that list is never empty
-	private @Builder.Default List<String> tags= new ArrayList<>();
-	
+
+	private @Singular List<Long> categories;
+
+	private @Builder.Default List<String> tags = new ArrayList<>();
+
 	private @Builder.Default List<String> references = new ArrayList<>();
-	
+
 	private Integer license;
 
 }
